Add files via upload
Very similar program that compares between two machine learning models (Random Forest and XGBoost) to study their differences and understand which fits best for the problem at hand. Works in the same way the other algorithm does, training and saving the model.

The biggest difference is that this algorithm predicts with both the models before showing the user any results, then plots the confusion matrix of the both of them at the same time.